feat: added ListGists command
With this command one can list all of his gists on github, choose one then edit all of its files in place inside neovim!

I added nvim_unception dependency, it is not strictly necessary but makes things look nicer;
with it the editable gist is in a neovim buffer in the already opened neovim session, not in a nested instance inside a terminal buffer.
Anyway the plugin works if without it